+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/generation/CodeAssembler.java
@@ -1,5 +1,8 @@
 package info.teksol.mc.mindcode.compiler.generation;
 
+import info.teksol.mc.common.SourcePosition;
+import info.teksol.mc.messages.ERR;
+import info.teksol.mc.messages.WARN;
 import info.teksol.mc.mindcode.compiler.CompilerMessageEmitter;
 import info.teksol.mc.mindcode.compiler.MindcodeInternalError;
 import info.teksol.mc.mindcode.compiler.ast.nodes.AstFunctionDeclaration;
@@ -8,20 +11,20 @@ import info.teksol.mc.mindcode.compiler.astcontext.AstContext;
 import info.teksol.mc.mindcode.compiler.astcontext.AstContextType;
 import info.teksol.mc.mindcode.compiler.astcontext.AstSubcontextType;
 import info.teksol.mc.mindcode.compiler.callgraph.MindcodeFunction;
-import info.teksol.mc.mindcode.compiler.generation.variables.NameCreator;
-import info.teksol.mc.mindcode.compiler.generation.variables.ValueStore;
-import info.teksol.mc.mindcode.compiler.generation.variables.Variables;
+import info.teksol.mc.mindcode.compiler.generation.variables.*;
 import info.teksol.mc.mindcode.logic.arguments.*;
 import info.teksol.mc.mindcode.logic.instructions.ContextfulInstructionCreator;
 import info.teksol.mc.mindcode.logic.instructions.CustomInstruction;
 import info.teksol.mc.mindcode.logic.instructions.InstructionProcessor;
 import info.teksol.mc.mindcode.logic.instructions.LogicInstruction;
 import info.teksol.mc.mindcode.logic.opcodes.InstructionParameterType;
+import info.teksol.mc.mindcode.logic.opcodes.NamedParameter;
 import info.teksol.mc.mindcode.logic.opcodes.Opcode;
 import info.teksol.mc.profile.GlobalCompilerProfile;
 import org.jspecify.annotations.NullMarked;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
 import java.util.function.Consumer;
@@ -281,5 +284,48 @@ public class CodeAssembler extends CompilerMessageEmitter implements ContextfulI
 
     public LogicInstruction createCustomInstruction(boolean safe, boolean text, boolean label, String opcode, List<LogicArgument> args, List<InstructionParameterType> params) {
         return addInstruction(new CustomInstruction(astContext, safe, text, label, opcode, args, params));
+    }
+
+    public LogicKeyword validateKeyword(NamedParameter parameter, FunctionArgument argument, boolean requireValidKeyword) {
+        if (argument.hasInModifier() || argument.hasOutModifier()) {
+            error(argument, ERR.ARGUMENT_KEYWORD_IN_OUT_NOT_ALLOWED);
+            return LogicKeyword.INVALID;
+        }
+
+        Collection<String> parameterValues = processor.getParameterValues(parameter.type());
+
+        LogicKeyword keyword = toKeyword(argument, requireValidKeyword);
+        if (!keyword.getKeyword().isEmpty()) {
+            if (!parameterValues.contains(keyword.getKeyword())) {
+                error(argument, ERR.ARGUMENT_KEYWORD_INVALID_VALUE, keyword.getKeywordLiteral(),
+                        ":" + String.join("', ':", parameterValues));
+                // Fill in one of the valid keywords to avoid further errors
+                return requireValidKeyword ? LogicKeyword.create(SourcePosition.EMPTY,
+                        parameterValues.iterator().next()) : LogicKeyword.INVALID;
+            }
+
+            return keyword;
+        } else {
+            error(argument, ERR.ARGUMENT_KEYWORD_UNSPECIFIED_VALUE,
+                    ":" + String.join("', ':", parameterValues));
+            // Fill in one of the valid keywords to avoid further errors
+            return requireValidKeyword ? LogicKeyword.create(SourcePosition.EMPTY,
+                    parameterValues.iterator().next()) : LogicKeyword.INVALID;
+        }
+    }
+
+    private LogicKeyword toKeyword(FunctionArgument argument, boolean requireValidKeyword) {
+        if (argument.unwrap() instanceof LogicKeyword keyword) {
+            return keyword;
+        } else if (argument instanceof IdentifierFunctionArgument ifa) {
+            if (requireValidKeyword) {
+                warn(argument.sourcePosition(), WARN.MISSING_MLOG_KEYWORD_PREFIX, ifa.getIdentifier().getName());
+            }
+            return ifa.getKeyword();
+        } else if (argument.unwrap() instanceof LogicBoolean bool) {
+            return LogicKeyword.create(argument.unwrap().sourcePosition(), bool.toMlog());
+        } else {
+            return LogicKeyword.INVALID;
+        }
     }
 }