+++ compiler/src/main/java/info/teksol/mc/mindcode/logic/opcodes/MindustryOpcodeVariants.java
@@ -370,6 +370,8 @@ public class MindustryOpcodeVariants {
                 type("type"), in("multiple"), in("min"), cond("minOp"), in("value"), cond("maxOp"), in("max"), in("message"));
         add(list, V6, MAX,  S, NONE, Opcode.ERROR,
                 in("e0"), in("e1"), in("e2"), in("e3"), in("e4"), in("e5"), in("e6"), in("e7"), in("e8"), in("e9"));
+        add(list, V8A,MAX,  S, NONE, Opcode.LOG, loglevel("level"),
+                in("e0"), in("e1"), in("e2"), in("e3"), in("e4"), in("e5"), in("e6"), in("e7"), in("e8"), in("e9"));
 
         // Virtual instructions
         add(list, V6,  MAX, S, NONE, Opcode.EMPTY);
@@ -464,6 +466,10 @@ public class MindustryOpcodeVariants {
 
     public static NamedParameter locate(String name) {
         return new NamedParameter(InstructionParameterType.LOCATE, name);
+    }
+
+    public static NamedParameter loglevel(String name) {
+        return new NamedParameter(InstructionParameterType.LOG_LEVEL, name);
     }
 
     public static NamedParameter lookup(String name) {